requirements.txt
pydantic
pymongo
spotipy
docker
python-dotenv
tqdm
# librosa : will be later usefull